==> 3_analyze_transactions.ipynb <==
# preprocess datasets

import pandas as pd
import plotly.express as px
import pickle
import re

exchange_rates = {'AED': 0.272234775270193, 'AFN': 0.013149243918474688, 'ALL': 0.009081035528643401, 'AMD': 0.0020537229273109416, 'ANG': 0.5437146585471945, 'AOA': 0.003156027697299072, 'ARS': 0.02437538086532602, 'AUD': 0.7092701610043266, 'AWG': 0.5552470849528041, 'AZN': 0.5865102639296187, 'BAM': 0.5812263876780006, 'BBD': 0.5011777677542224, 'BDT': 0.011871972646975022, 'BGN': 0.5780346820809249, 'BHD': 2.6518164942985947, 'BIF': 0.0005525014503163071, 'BMD': 1.0, 'BND': 0.7404116688879017, 'BOB': 0.14484147100997957, 'BRL': 0.263664407941572, 'BSD': 0.9995002498750625, 'BTN': 0.014491598495772076, 'BWP': 0.09384384384384384, 'BYR': 5.102040816326531e-05, 'BZD': 0.49512303807496166, 'CAD': 0.7476635514018692, 'CDF': 0.0006134969325153375, 'CHF': 1.0036130068245686, 'CLP': 0.0014983437308399297, 'CNY': 0.14896247635220689, 'COP': 0.0003237293622531564, 'CRC': 0.0016714162745802656, 'CUP': 0.03773584905660377, 'CVE': 0.01028843631200712, 'CZK': 0.043991043423558966, 'DJF': 0.005626828719333784, 'DKK': 0.15150826477584353, 'DOP': 0.01971803214039239, 'DZD': 0.008415737428992215, 'EGP': 0.05796429399489915, 'ERN': 0.0666680000266672, 'ETB': 0.0350594257266066, 'EUR': 1.075268817204301, 'FJD': 0.47107593744111553, 'FKP': 1.3157894736842106, 'GBP': 1.25, 'GEL': 0.37243947858473, 'GHS': 0.191131498470948, 'GIP': 1.3157894736842106, 'GMD': 0.02013490385583409, 'GNF': 0.00010937445261504419, 'GTQ': 0.12978248455588434, 'GYD': 0.0047971984361133094, 'HKD': 0.1274356131564527, 'HNL': 0.04082299150881777, 'HRK': 0.1523972080831479, 'HTG': 0.012072556061932213, 'HUF': 0.0035799052757064047, 'IDR': 7.048433309486134e-05, 'ILS': 0.2768166089965398, 'INR': 0.014496955639315743, 'IQD': 0.0008361204013377926, 'IRR': 2.3750148382020786e-05, 'ISK': 0.008419634587858887, 'JMD': 0.00799998720002048, 'JOD': 1.4104372355430184, 'JPY': 0.009048185205493334, 'KES': 0.009930595071048444, 'KGS': 0.014335622732283679, 'KHR': 0.0002497533560732099, 'KMF': 0.0023110668207994164, 'KPW': 0.001111005812449109, 'KRW': 0.0008821474786504873, 'KWD': 3.2948929159802307, 'KYD': 1.1971746677840296, 'KZT': 0.002642145443237072, 'LAK': 0.00011620749405860134, 'LBP': 0.0006605675583249778, 'LKR': 0.005613240961840065, 'LRD': 0.006168080185042405, 'LSL': 0.06954006202973533, 'LYD': 0.7204610951008646, 'MAD': 0.1039619914959091, 'MDL': 0.05778008898133703, 'MGA': 0.00028064572765150713, 'MKD': 0.01847762821164275, 'MMK': 0.0006520818135751566, 'MNT': 0.0003800945112604139, 'MOP': 0.12343088487601368, 'MUR': 0.02896326010455737, 'MVR': 0.06454944487477407, 'MWK': 0.0013812440865487545, 'MXN': 0.05270786664909738, 'MYR': 0.2462690242821258, 'MZN': 0.01581525254586028, 'NAD': 0.0695405456151209, 'NGN': 0.0027636524430687598, 'NIO': 0.03034394865803887, 'NOK': 0.11728554338392251, 'NPR': 0.009094629621208676, 'NZD': 0.6869075422448139, 'OMR': 2.5974025974025974, 'PAB': 0.9995002498750625, 'PEN': 0.30414550320873507, 'PGK': 0.2962524070508073, 'PHP': 0.019041997124658434, 'PKR': 0.007134184013435096, 'PLN': 0.2635323881305012, 'PYG': 0.00016226521724018602, 'QAR': 0.274649821477616, 'RON': 0.2377555872563005, 'RSD': 0.009584061322657968, 'RUB': 0.009991666949763897, 'RWF': 0.0011065190570244596, 'SAR': 0.26666666666666666, 'SBD': 0.12413570515287313, 'SCR': 0.0731288164101064, 'SDG': 0.02095210558185045, 'SEK': 0.10798319781442008, 'SGD': 0.7401924500370096, 'SHP': 0.7570595805889924, 'SLL': 0.0001126760525292602, 'SOS': 0.0017182136488706696, 'SRD': 0.13408420488066505, 'SVC': 0.11401468509143979, 'SYP': 0.0019417468187390998, 'SZL': 0.07008199593524424, 'THB': 0.031555695803092455, 'TJS': 0.10569372073605107, 'TMT': 0.2857142857142857, 'TND': 0.3311587243765937, 'TOP': 0.4430267588162325, 'TRY': 0.18021914648212226, 'TTD': 0.1472515498225619, 'TWD': 0.03243909559801472, 'TZS': 0.0004266024146549874, 'UAH': 0.03677133012932477, 'UGX': 0.0002696506204323713, 'USD': 1.0, 'UYU': 0.030084235860409144, 'UZS': 0.00011910288562233081, 'VEF': 0.10012515644555695, 'VND': 4.310066159515549e-05, 'VUV': 0.008782808179956227, 'WST': 0.3848077885096394, 'XAF': 0.0017329205516579283, 'XCD': 0.36991824806717716, 'XDR': 1.3914011409489355, 'XOF': 0.0017329214525624882, 'XPF': 0.009531997963965234, 'YER': 0.00399440144693198, 'ZAR': 0.06973646589538135, 'ZMW': 0.08313242996092776, 'ZWL': 0.0031021699678925405}

# load fig_dict
fig_dict = pd.read_pickle('fig_dict.pkl')

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

# Load data from csv
active_users_df = pd.read_pickle('active_users_df.pkl')
all_collectives = pd.read_csv('data/all_collectives.csv', low_memory=False)
# expenses aggregated by collective, including child collectives
# Load data from csv, sql query that generated this csv is queries/expenses_in_period.sql
expenses_df = pd.read_csv('data/expenses_in_period_by_collective.csv', low_memory=False)
# sql query that generated this csv is queries/income_in_period.sql
income_df = pd.read_csv('data/income_in_period_by_collective.csv', low_memory=False)

def camel_to_snake(name):
    s1 = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', name)
    return re.sub('([a-z0-9])([A-Z])', r'\1_\2', s1).lower().replace('_i_d', '_id')

expenses_df.columns = [camel_to_snake(col) for col in expenses_df.columns]
income_df.columns = [camel_to_snake(col) for col in income_df.columns]
all_collectives.columns = [camel_to_snake(col) for col in all_collectives.columns]

# drop expenses and income rows where amount is 0
expenses_df = expenses_df[expenses_df['total_amount'] != 0]
income_df = income_df[income_df['total_amount'] != 0]

# add a new column to expenses_df amountInUSD that is the product of (amountInHostCurrency/100 * exchange) rate for hostCurrency, but if hostCurrency is USD, then amountInUSD is just amountInHostCurrency/100
expenses_df['amount_in_usd'] = expenses_df.apply(lambda row: row['total_amount']/100 * exchange_rates[row['currency']] if row['currency'] != 'USD' else row['total_amount']/100, axis=1)
income_df['amount_in_usd'] = income_df.apply(lambda row: row['total_amount']/100 * exchange_rates[row['currency']] if row['currency'] != 'USD' else row['total_amount']/100, axis=1)
income_df['platform_tip_amount_in_usd'] = income_df.apply(lambda row: row['platform_tip_amount']/100 * exchange_rates[row['currency']] if row['currency'] != 'USD' else row['platform_tip_amount']/100, axis=1)

# add collective slug to expenses_df by merging with all_collectives on collective_id
expenses_df = pd.merge(expenses_df, all_collectives[['id', 'slug']], left_on='collective_id', right_on='id', how='left')
income_df = pd.merge(income_df, all_collectives[['id', 'slug']], left_on='collective_id', right_on='id', how='left')

# rename slug to collective_slug
expenses_df.rename(columns={'slug': 'collective_slug'}, inplace=True)
income_df.rename(columns={'slug': 'collective_slug'}, inplace=True)

# drop id column
expenses_df.drop('id', axis=1, inplace=True)
income_df.drop('id', axis=1, inplace=True)

# add host collective slug to expenses_df by merging with all_collectives on host_collective_id
expenses_df = pd.merge(expenses_df, all_collectives[['id', 'host_fee_percent', 'slug']], left_on='host_collective_id', right_on='id', how='left')
income_df = pd.merge(income_df, all_collectives[['id', 'host_fee_percent', 'slug']], left_on='host_collective_id', right_on='id', how='left')

# rename slug to host_collective_slug
expenses_df.rename(columns={'slug': 'host_collective_slug'}, inplace=True)
income_df.rename(columns={'slug': 'host_collective_slug'}, inplace=True)

# drop id column
expenses_df.drop('id', axis=1, inplace=True)
income_df.drop('id', axis=1, inplace=True)

# reorder columns in expenses_df like this: collective_id, currency, transaction_count, total_amount, amount_in_usd
expenses_df = expenses_df[['collective_id', 'collective_slug', 'host_collective_id', 'host_collective_slug', 'host_fee_percent', 'transaction_count', 'amount_in_usd']]
income_df = income_df[['collective_id', 'collective_slug', 'host_collective_id', 'host_collective_slug', 'payment_method_type', 'host_fee_percent', 'transaction_count', 'amount_in_usd', 'platform_tip_amount_in_usd']]

# merge rows in expenses_df where 'collective_id', 'collective_slug', 'host_collective_id', 'host_fee_percent', 'host_collective_slug' are the same, and sum 'transaction_count' and 'amount_in_usd'
expenses_df = expenses_df.groupby(['collective_id', 'collective_slug', 'host_collective_id', 'host_collective_slug', 'host_fee_percent', ]).agg({'transaction_count': 'sum', 'amount_in_usd': 'sum'}).reset_index()
income_df = income_df.groupby(['collective_id', 'collective_slug', 'host_collective_id', 'host_collective_slug', 'payment_method_type', 'host_fee_percent']).agg({'transaction_count': 'sum', 'amount_in_usd': 'sum' , 'platform_tip_amount_in_usd': 'sum'}).reset_index()

# format amount_in_usd as int
expenses_df['amount_in_usd'] = abs(expenses_df['amount_in_usd'].astype(int))
income_df['amount_in_usd'] = abs(income_df['amount_in_usd'].astype(int))
income_df['platform_tip_amount_in_usd'] = abs(income_df['platform_tip_amount_in_usd'].astype(int))

# if transaction_count for expenses is less than 5 and collective_id is same as host_collective_id, then drop that row and save the collective_ids that were dropped in a list
# Getting the collective_id values of rows that meet the drop condition
dropped_collective_ids = expenses_df.loc[(expenses_df['transaction_count'] < 5) & (expenses_df['collective_id'] == expenses_df['host_collective_id']), 'collective_id']

# Apply the filter to retain the rows in expenses_df that don't meet the drop condition
expenses_df = expenses_df[~((expenses_df['transaction_count'] < 5) & (expenses_df['collective_id'] == expenses_df['host_collective_id']))]

# Drop same rows from income_df
income_df = income_df[~income_df['collective_id'].isin(dropped_collective_ids)]

# drop collectives with slug ending with <string>-<number>.<number> as this is a known spammer pattern for some reason
expenses_df = expenses_df[~expenses_df['collective_slug'].str.contains(r'-\d+\.\d+$')]
income_df = income_df[~income_df['collective_slug'].str.contains(r'-\d+\.\d+$')]

# drop internal opencollective collectives where host_collective_slug is 'opencollective'
expenses_df = expenses_df[~(expenses_df['host_collective_slug'] == 'opencollective')]
income_df = income_df[~(income_df['host_collective_slug'] == 'opencollective')]

# Drop rows from income_df if the sum of all transaction_count for rows with that collective_id in income_df is less than 5.

# 1. Group by 'collective_id' and get sum of 'transaction_count' for each 'collective_id'
transaction_sums = income_df.groupby('collective_id')['transaction_count'].sum()

# 2. Filter out 'collective_id's where the sum of 'transaction_count' is less than 5
low_transaction_ids = transaction_sums[transaction_sums < 5].index

# 3. Group by 'collective_id' and check if all 'payment_method_type' for each 'collective_id' is 'host'
all_host = income_df.groupby('collective_id')['payment_method_type'].apply(lambda x: all((val == 'host' or val == 'blank') for val in x))

# 4. Filter out 'collective_id's where all 'payment_method_type' is 'host'
host_payment_ids = all_host[all_host].index

# 5. Combine 'collective_id's based on the conditions and their absence in 'expenses_df'
dropped_collective_ids_from_low_transactions = set(low_transaction_ids) - set(expenses_df['collective_id'])
dropped_collective_ids_from_host_payments = set(host_payment_ids) - set(expenses_df['collective_id'])
dropped_collective_ids = dropped_collective_ids_from_low_transactions.union(dropped_collective_ids_from_host_payments)

# 6. Drop rows from 'income_df' that have 'collective_id's in 'dropped_collective_ids'
income_df = income_df[~income_df['collective_id'].isin(dropped_collective_ids)]


# Expenses by collective, host and host fee percent

# Expense amount by host and collective

expense_tree_map = expenses_df.copy()

# Aggregate data based on host_collective_slug and collective_slug
aggregated_expenses = expense_tree_map.groupby(['host_collective_slug', 'collective_slug']).agg({
    'amount_in_usd': 'sum',
    'transaction_count': 'sum'
}).reset_index()

# Plot treemap with aggregated data
fig = px.treemap(aggregated_expenses, 
                 path=['host_collective_slug', 'collective_slug'], 
                 values='amount_in_usd', 
                 color='host_collective_slug', 
                 color_discrete_sequence=px.colors.qualitative.Pastel,
                 custom_data=['collective_slug', 'host_collective_slug', 'transaction_count', 'amount_in_usd'])
fig.update_layout(title='Expense <b>amount</b> in USD, last 12 months, for all hosts and collectives <br><i>Box size is proportional to expense amount in USD</i>')
# Define hovertemplates
hovertemplate = (
    '<b>%{label}</b><br>'
    '<b>Host:</b> %{customdata[1]}<br>'
    '<b>Amount in USD:</b> $%{customdata[3]:,}<br>'
    '<extra></extra>'  # To hide extra info
)

# Update hovertemplate
fig.update_traces(hovertemplate=hovertemplate)

fig.show()

fig_expense_amount_by_host_and_collective = fig


# Expense count by host and collective

expense_tree_map = expenses_df.copy()

# Aggregate data based on host_collective_slug and collective_slug
aggregated_expenses = expense_tree_map.groupby(['host_collective_slug', 'collective_slug']).agg({
    'transaction_count': 'sum',
    'amount_in_usd': 'sum' 
}).reset_index()

# Plot treemap with aggregated data
fig = px.treemap(aggregated_expenses, 
                 path=['host_collective_slug', 'collective_slug'], 
                 values='transaction_count', 
                 color='host_collective_slug', 
                 color_discrete_sequence=px.colors.qualitative.Pastel,
                 custom_data=['collective_slug', 'host_collective_slug', 'transaction_count', 'amount_in_usd'])
fig.update_layout(title='Expense <b>count</b>, last 12 months, for all hosts and collectives <br><i>Box size is proportional to number of expenses</i>')

# Define hovertemplates
hovertemplate = (
    '<b>%{label}</b><br>'
    '<b>Host:</b> %{customdata[1]}<br>'
    '<b>Expense count:</b> %{customdata[2]:,}<br>'
    '<extra></extra>'  # To hide extra info
)

# Update hovertemplate
fig.update_traces(hovertemplate=hovertemplate)

fig.show()

fig_expense_count_by_host_and_collective = fig


# for expenses_df, plot a tree map of collectives by transaction_count with plotly where the size of the square is proportional to transaction_count
# color the squares by host_fee_percent

# Prepare data
expense_tree_map = expenses_df.copy()
expense_tree_map['host_fee_percent'] = expense_tree_map['host_fee_percent'].apply(lambda x: str(int(x)) + '%')

# Plot treemap
fig = px.treemap(expense_tree_map, 
                 path=['host_fee_percent', 'host_collective_slug', 'collective_slug'], 
                 values='transaction_count', 
                 color='host_fee_percent', 
                 color_continuous_scale=px.colors.qualitative.Pastel, 
                 custom_data=['collective_slug', 'host_collective_slug', 'transaction_count', 'amount_in_usd'])

fig.update_layout(title='Expense <b>count</b> by host fee %, last 12 months, for all collectives <br><i>Box size is proportional to number of expenses</i>')

# Define hovertemplates
hovertemplate = (
    '<b>%{label}</b><br>'
    '<b>Host:</b> %{customdata[1]}<br>'
    '<b>Expense count:</b> %{customdata[2]:,}<br>'
    '<extra></extra>'  # To hide extra info
)

# Update hovertemplate
fig.update_traces(hovertemplate=hovertemplate)

fig.show()

fig_expense_count_by_host_fee = fig

expense_tree_map.head()

# for expenses_df, plot a tree map of collectives by amount_in_usd with plotly where the size of the square is proportional to amount_in_usd
# color the squares by host_fee_percent

expense_tree_map = expenses_df.copy()

expense_tree_map['host_fee_percent'] = expense_tree_map['host_fee_percent'].apply(lambda x: str(int(x)) + '%')

fig = px.treemap(expense_tree_map, path=['host_fee_percent', 'host_collective_slug', 'collective_slug'], values='amount_in_usd', color='host_fee_percent', color_continuous_scale=px.colors.qualitative.Pastel, custom_data=['collective_slug', 'host_collective_slug', 'transaction_count', 'amount_in_usd'])
fig.update_layout(title='Expense <b>amount</b> in USD by host fee %, last 12 months, for all collectives <br><i>Box size is proportional to expense amount in USD</i>')
# Define hovertemplates
hovertemplate = (
    '<b>%{label}</b><br>'
    '<b>Host:</b> %{customdata[1]}<br>'
    '<b>Amount in USD:</b> $%{customdata[3]:,}<br>'
    '<extra></extra>'  # To hide extra info
)

# Update hovertemplate
fig.update_traces(hovertemplate=hovertemplate)

fig.show()

fig_expense_amount_by_host_fee = fig


# for expenses_df with host_fee_percent = 0, plot a tree map of collectives by amount_in_usd with plotly where the size of the square is proportional to amount_in_usd
# color the squares by collective_host_slug

# Filter data where host_fee_percent = 0
expenses_df_host_fee_0 = expenses_df[expenses_df['host_fee_percent'] == 0].copy()

# Aggregate data based on host_collective_slug and collective_slug
aggregated_expenses = expenses_df_host_fee_0.groupby(['host_collective_slug', 'collective_slug']).agg({
    'amount_in_usd': 'sum',
    'transaction_count': 'sum'
}).reset_index()

# Plot treemap with aggregated data
fig = px.treemap(aggregated_expenses, 
                 path=['host_collective_slug', 'collective_slug'], 
                 values='amount_in_usd', 
                 color='host_collective_slug', 
                 color_discrete_sequence=px.colors.qualitative.Pastel,
                 custom_data=['collective_slug', 'host_collective_slug', 'transaction_count', 'amount_in_usd'])
fig.update_layout(title='Expense <b>amount</b> in USD, last 12 months, for collectives with no host fee <br><i>Box size is proportional to expense amount in USD</i>')
# Define hovertemplates
hovertemplate = (
    '<b>%{label}</b><br>'
    '<b>Host:</b> %{customdata[1]}<br>'
    '<b>Amount in USD:</b> $%{customdata[3]:,}<br>'
    '<extra></extra>'  # To hide extra info
)

# Update hovertemplate
fig.update_traces(hovertemplate=hovertemplate)

fig.show()

fig_expense_amount_by_host_fee_for_zero_hosts = fig


# for expenses_df with host_fee_percent = 0, plot a tree map of collectives by transaction_count with plotly where the size of the square is proportional to transaction_count
# color the squares by collective_host_slug

# Filter data where host_fee_percent = 0
expenses_df_host_fee_0 = expenses_df[expenses_df['host_fee_percent'] == 0].copy()

# Aggregate data based on host_collective_slug and collective_slug
aggregated_expenses = expenses_df_host_fee_0.groupby(['host_collective_slug', 'collective_slug']).agg({
    'transaction_count': 'sum',
    'amount_in_usd': 'sum' 
}).reset_index()

# Plot treemap with aggregated data
fig = px.treemap(aggregated_expenses, 
                 path=['host_collective_slug', 'collective_slug'], 
                 values='transaction_count', 
                 color='host_collective_slug', 
                 color_discrete_sequence=px.colors.qualitative.Pastel,
                 custom_data=['collective_slug', 'host_collective_slug', 'transaction_count', 'amount_in_usd'])
fig.update_layout(title='Expense <b>count</b>, last 12 months, for collectives with no host fee <br><i>Box size is proportional to number of expenses</i>')

# Define hovertemplates
hovertemplate = (
    '<b>%{label}</b><br>'
    '<b>Host:</b> %{customdata[1]}<br>'
    '<b>Expense count:</b> %{customdata[2]:,}<br>'
    '<extra></extra>'  # To hide extra info
)

# Update hovertemplate
fig.update_traces(hovertemplate=hovertemplate)

fig.show()

fig_expense_count_by_host_fee_for_zero_hosts = fig


# expense by collective

# display top 100 rows of expenses_df sorted by amount_in_usd
expenses_df.sort_values(by='amount_in_usd').head(10)

# Income by collective, host and host fee percent - vs payment method type

# income by host

income_df.groupby(['host_collective_slug']).agg({'transaction_count': 'sum', 'amount_in_usd': 'sum'}).reset_index().sort_values('amount_in_usd', ascending=False).head(10)

# for income_df, plot a tree map of payment_method_type by amount_in_usd with plotly where the size of the square is proportional to amount_in_usd
# color the squares by host_collective_slug

income_tree_map = income_df.copy()

income_tree_map_aggregated = income_tree_map.groupby(['host_collective_slug', 'payment_method_type']).agg({'amount_in_usd': 'sum'}).reset_index()

fig = px.treemap(income_tree_map_aggregated, 
                 path=['host_collective_slug', 'payment_method_type'], 
                 values='amount_in_usd', 
                 color='host_collective_slug', 
                 color_discrete_sequence=px.colors.qualitative.Pastel, 
                 custom_data=['amount_in_usd'])

fig.update_layout(title='Income <b>amount</b> in USD, last 12 months, by host and payment method type <br><i>Box size is proportional to payment method type income amount in USD</i>')
fig.update_traces(hovertemplate='<b>%{label}</b><br>Amount in USD: $%{customdata[0]:,.2f}')

fig.show()

fig_income_amount_by_host_and_method = fig


# for income_df, plot a tree map of payment_method_type by amount_in_usd with plotly where the size of the square is proportional to amount_in_usd
# color the squares by host_fee_percent

income_tree_map = income_df.copy()
income_tree_map['amount_in_usd'] = income_tree_map['amount_in_usd'].abs()
income_tree_map['host_fee_percent'] = income_tree_map['host_fee_percent'].apply(lambda x: str(int(x)) + '%')

income_tree_map['tip_possible'] = income_tree_map['payment_method_type'].apply(lambda x: 'Tip possible' if x not in ['host', 'blank', 'collective'] else 'No tip possible')

income_tree_map_aggregated = income_tree_map.groupby(['host_fee_percent', 'host_collective_slug', 'tip_possible', 'payment_method_type']).agg({'amount_in_usd': 'sum'}).reset_index()

fig = px.treemap(income_tree_map_aggregated, 
                 path=['host_fee_percent', 'host_collective_slug', 'tip_possible', 'payment_method_type'], 
                 values='amount_in_usd', 
                 color='tip_possible', 
                 color_continuous_scale=px.colors.qualitative.Alphabet_r, 
                 custom_data=['amount_in_usd'])

fig.update_layout(title='Income <b>amount</b> in USD by host fee % and payment method, last 12 months, for all collectives <br><i>Box size is proportional to payment method type income amount in USD</i>')
fig.update_traces(hovertemplate='<b>%{label}</b><br>Amount in USD: $%{customdata[0]:,.2f}')

fig.show()

fig_income_amount_by_host_percentage_and_method = fig

# collectives and hosts by platform tip amount

income_tree_map = income_df.copy()

income_tree_map['host_fee'] = income_tree_map['host_fee_percent'].apply(lambda x: 'Host fee' if x > 0 else 'No host fee')

income_tree_map_aggregated = income_tree_map.groupby(['host_fee', 'host_collective_slug', 'collective_slug']).agg({'platform_tip_amount_in_usd': 'sum'}).reset_index()

fig = px.treemap(income_tree_map_aggregated, 
                 path=['host_fee', 'host_collective_slug', 'collective_slug'], 
                 values='platform_tip_amount_in_usd', 
                 color_continuous_scale=px.colors.qualitative.Antique, 
                 custom_data=['platform_tip_amount_in_usd'])

fig.update_layout(title='Income from platform tips in USD per host and collective <br><i>Box size is proportional to platform tip income in USD</i>')
fig.update_traces(hovertemplate='<b>%{label}</b><br>Platform tip amount: $%{customdata[0]:,.2f}')

fig.show()

fig_tip_amount_amount_by_host_fee_and_host_and_collective = fig

# for income_df, plot a tree map of payment_method_type by amount_in_usd with plotly where the size of the square is proportional to amount_in_usd
# color the squares by host_collective_slug

# get only rows where host_fee_percent is 0
income_tree_map = income_df[income_df['host_fee_percent'] == 0].copy()
income_tree_map_aggregated = income_tree_map.groupby(['payment_method_type', 'host_collective_slug']).agg({'transaction_count': 'sum', 'amount_in_usd': 'sum'}).reset_index()

fig = px.treemap(income_tree_map_aggregated, 
                 path=['payment_method_type', 'host_collective_slug'], 
                 values='amount_in_usd', 
                 color='payment_method_type', 
                 color_continuous_scale=px.colors.qualitative.Pastel, 
                 custom_data=['amount_in_usd'])

fig.update_layout(title='Income <b>amount</b> in USD by payment method type, last 12 months, for collectives with no host fee <br><i>Box size is proportional to collective income amount in USD</i>')

# Define hovertemplates
hovertemplate = (
    '<b>%{label}</b><br>'
    '<b>Amount in USD:</b> $%{customdata[0]:,}<br>'
    '<extra></extra>'  # To hide extra info
)

# Update hovertemplate
fig.update_traces(hovertemplate=hovertemplate)

fig.show()

fig_income_amount_method_and_host_for_zero_hosts = fig

treemaps = [
    fig_expense_amount_by_host_and_collective,
    fig_expense_count_by_host_and_collective,
    fig_expense_count_by_host_fee,
    fig_expense_amount_by_host_fee,
    fig_expense_amount_by_host_fee_for_zero_hosts,
    fig_expense_count_by_host_fee_for_zero_hosts,
    fig_income_amount_by_host_and_method,
    fig_income_amount_by_host_percentage_and_method,
    fig_tip_amount_amount_by_host_fee_and_host_and_collective,
    fig_income_amount_method_and_host_for_zero_hosts
]

fig_dict['treemaps'] = {}

fig_dict['treemaps']['expense_amount_by_host_and_collective'] = {
    'fig': fig_expense_amount_by_host_and_collective,
    'description': 'Expense amount by host and collective'
}

fig_dict['treemaps']['expense_count_by_host_and_collective'] = {
    'fig': fig_expense_count_by_host_and_collective,
    'description': 'Expense count by host and collective'
}

fig_dict['treemaps']['expense_count_by_host_fee'] = {
    'fig': fig_expense_count_by_host_fee,
    'description': 'Expense count by host fee'
}

fig_dict['treemaps']['expense_amount_by_host_fee'] = {
    'fig': fig_expense_amount_by_host_fee,
    'description': 'Expense amount by host fee'
}

fig_dict['treemaps']['expense_amount_by_host_fee_for_zero_hosts'] = {
    'fig': fig_expense_amount_by_host_fee_for_zero_hosts,
    'description': 'Expense amount by host fee for zero hosts'
}

fig_dict['treemaps']['expense_count_by_host_fee_for_zero_hosts'] = {
    'fig': fig_expense_count_by_host_fee_for_zero_hosts,
    'description': 'Expense count by host fee for zero hosts'
}

fig_dict['treemaps']['income_amount_by_host_and_method'] = {
    'fig': fig_income_amount_by_host_and_method,
    'description': 'Income amount by host and method'
}

fig_dict['treemaps']['income_amount_by_host_percentage_and_method'] = {
    'fig': fig_income_amount_by_host_percentage_and_method,
    'description': 'Income amount by host percentage and method'
}

fig_dict['treemaps']['tip_amount_amount_by_host_fee_and_host_and_collective'] = {
    'fig': fig_tip_amount_amount_by_host_fee_and_host_and_collective,
    'description': 'Tip amount amount by host fee and host and collective'
}

fig_dict['treemaps']['income_amount_method_and_host_for_zero_hosts'] = {
    'fig': fig_income_amount_method_and_host_for_zero_hosts,
    'description': 'Income amount method and host for zero hosts'
}
    


# save fig_dict to pkl file
with open('fig_dict.pkl', 'wb') as f:
    pickle.dump(fig_dict, f)
