==> 4_generate_report.ipynb <==
import pandas as pd
import plotly.offline as offline
import pickle

# load fig_dict
fig_dict = pd.read_pickle('fig_dict.pkl')

fig_dict['user_clusters'][0]['info']

def set_fig_height(fig, height):
    fig.layout.height = height
    return fig

def get_html_section(data, include_figs=False, header_level='h1', fig_height=500):
    if include_figs and data['figs']:
        if not isinstance(data['figs'], list):
            data['figs'] = [data['figs']]
        figs = [offline.plot(set_fig_height(fig['fig'], fig_height), include_plotlyjs=False, output_type='div', show_link=False) for fig in data['figs']]
        figs = ''.join(figs)
        return f"""
        <div class='chart-container'>
            <{header_level} id="{data['id']}">{data['title']}</{header_level}>
            <p>{data['body']}</p>
            {figs}
        </div>
        """
    else:
        return f"""
        <div class='chart-container'>
            <{header_level} id="{data['id']}">{data['title']}</{header_level}>
            <p>{data['body']}</p>
        </div>
        """

def get_html_figs_section(figs, fig_info, header_level='h1', fig_height=500):
    if not isinstance(figs, list):
        figs = [figs]
    
    fig_divs = []

    for a in figs:
        fig = a['fig']
        fig.layout.height = fig_height
        div = offline.plot(fig, include_plotlyjs=False, output_type='div', show_link=False)
        details_section = f"""
        <details>
            <summary>More info</summary>
            <div class='info-div-container'>{a['info']}</div>
        </details>
        """ if 'info' in a else ""
        fig_divs.append(f"""
        <div class='chart-container'>
            <h2 id="{a['title'].replace(' ', '-').replace(':','-').lower()}">{a['title']}</h2>
            <p>{a['description']}</p>
            {div}
            {details_section}
        </div>
        """)

    all_divs = ''.join(fig_divs)
    
    # Generate the details section if the 'info' key exists and is not blank


    title_section = f"""
    <{header_level} id="{fig_info['title'].replace(' ', '-').replace(':','-').lower()}">{fig_info['title']}</{header_level}>
    """ if fig_info.get('title') else ""

    description_section = f"""
    <p>{fig_info['description']}</p>
    """ if fig_info.get('description') else ""

    return f"""
    <div class='chart-container'>
        {title_section}
        {description_section}
        {all_divs}
    </div>
    """

def generate_toc(header_sections):
    toc_items = []
    
    for _, section_data in header_sections.items():
        # Add the main section (h1) to TOC
        toc_items.append(f'<li><a href="#{section_data["id"]}">{section_data["title"]}</a></li>')
        
        # Add subsections (h2) to TOC if they exist
        if 'figs' in section_data and isinstance(section_data['figs'], list) and 'title' in section_data['figs'][0].keys():
            toc_items.append('<ul>')
            for fig in section_data['figs']:
                if 'title' in fig.keys():
                    fig_title = fig['title'].replace(' ', '-').replace(':','-').lower()
                    toc_items.append(f'<li><a href="#{fig_title}">{fig["title"]}</a></li>')
            toc_items.append('</ul>')

    return f"""
    <div class="side-toc">
        <h3>Table of Contents</h3>
        <ul>{''.join(toc_items)}</ul>
    </div>
    """


# Update the CSS styling for TOC and font
css_styles = """
<style>
    @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;500&display=swap');

    body {
        font-family: 'Roboto', sans-serif;
        font-size: 16px; /* A baseline font size */
        line-height: 1.6;
        color: #333; /* A slightly softer black for the text */
        background-color: #f5f5f5; /* A light gray background to contrast with white cards */
        margin: 0;
        padding: 0;
    }

    h1 {
        font-size: 2.5em;
        margin-bottom: 0.5em;
    }

    h2 {
        font-size: 2em;
        margin-bottom: 0.5em;
    }

    h3 {
        font-size: 1.5em;
        margin-bottom: 0.5em;
    }

    p {
        margin-bottom: 1em;
    }

    ul, ol {
        margin-bottom: 1em;
        padding-left: 2em;
    }

    .toc-container, .side-toc, .chart-container {
        background-color: #fff;
        border-radius: 8px;
        box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
        padding: 20px;
        margin-bottom: 2em;
    }

    .side-toc {
        position: fixed;
        top: 20px;
        left: 20px;
        width: 250px;
        max-height: 80vh;
        overflow-y: auto;
    }

    .toc-container a, .side-toc a {
        color: #0077cc; /* A neutral link color */
        text-decoration: none;
    }

    .toc-container ul, .side-toc ul {
        margin-bottom: 0;
    }

    .toc-container ul li, .side-toc ul li {
        margin-top: 1em;
    }

    .toc-container ul ul li, .side-toc ul ul li {
        margin-top: 0;
    }

    .toc-container a:hover, .side-toc a:hover {
        text-decoration: underline;
    }

    .info-div-container {
        margin-top: 10px;
    }

    details[open] summary ~ * {
        animation: fadeIn 0.3s ease-in;
    }

    @keyframes fadeIn {
        from { opacity: 0; }
        to { opacity: 1; }
    }

    details {
        cursor: pointer;
        margin-bottom: 0.25em;
    }

    summary {
        font-weight: 500;
        cursor: pointer;
        display: flex;
        align-items: center;
    }

    summary::-webkit-details-marker {
        display: none; /* hides the default down arrow for webkit browsers */
    }

    summary:before {
        content: '▶'; /* Right-pointing caret */
        display: inline-block;
        font-size: 0.8em;
        margin-right: 8px;
        transition: transform 0.3s ease-in-out;
    }

    details[open] summary:before {
        transform: rotate(90deg); /* Rotate the caret 90 degrees to point downwards when details is open */
    }
    
</style>

"""

sections = {
    'intro': {
        'id': 'intro',
        'title': 'Retreat Report: Users, Hosts and Collectives',
        'body': """

            <p>When we meet for the Open Collective team retreat in Bath, we'll delve into the past, present, and future of the platform and our hosts. As we strive to understand the best course of action, many 'what-ifs' and 'how-abouts' will arise. Much of our perception about Open Collective is shaped by assumptions about our users, collectives, and hosts. I've noticed that discussions often become mired in debates about these assumptions. However, debating about the present, which can be more accurately assessed than the future, seems redundant. This report is intended to serve as a guide as you ponder the future of Open Collective. I've structured it to address frequently revisited questions.</p>
            <details>
                <summary>How many users do we have?</summary>
                <div class="info-div-container">
                    <p>Defining 'user' is not straightforward. Someone who contributed five years ago and hasn't returned since is likely not considered a current user. A more pertinent metric would be the <a href="#active-users">roughly 100,000 people who have taken some important action</a> on the platform in the last 12 months. </p>
                </div>
            </details>

            <details>
                <summary>What do our users do on the platform, and how often do they do it?</summary>
                <div class="info-div-container">
                    <p>We don't track all clicks of users, but <a href="#user-activity">we count some key activities</a>.</p>
                </div>
            </details>

            <details>
                <summary>Can we categorize our users? What are these categories and how many users belong to each?</summary>
                <div class="info-div-container">
                    <p>Categorization can be approached in myriad ways. A practical method is to group users by their actions to discern patterns. <a href="#user-cluster-intro">In this report, we identified 13 such clusters</a>. The subsequent phase involves using these clusters to craft user profiles, incorporating qualitative data to understand their motivations.</p>
                    <p>Notably, our largest user segment consists of one-time contributors. They represent a more than half of the users who engaged with the platform in the past 12 months.</p>
                </div>
            </details>

            <details>
                <summary>Who uses the platform, besides the collectives of the three primary hosts?</summary>
                <div class="info-div-container">
                    <p><a href="#collective-expenses">Many independent hosts and collectives exist</a>. Approximately a third of the expenses submitted are linked to collectives not managed by the primary hosts.</p>
                </div>
            </details>

            <details>
                <summary>How many active hosts are there? What's their size range? How many independent collectives?</summary>
                <div class="info-div-container">
                    <p>The vast disparities in the sizes of independent hosts and collectives make a mere count less insightful. <a href="#collective-expenses">A graphical representation of hosts and collectives</a>, segregated by the number and amount of submitted expenses, might offer a clearer perspective.</p>
                </div>
            </details>

            <details>
                <summary>How much money is being disbursed via the platform? What's the distribution among the hosts?</summary>
                <div class="info-div-container">
                    <p>The platform facilitated payouts totaling around <a href="#collective-expenses">45 million USD in the last year</a>. Of this, approximately 27 million was channeled through primary hosts. A breakdown by host is provided for further clarity.</p>
                </div>
            </details>

            <details>
                <summary>How many expenses are processed, and who handles them?</summary>
                <div class="info-div-container">
                    <p>Over the past year, <a href="#collective-expenses">about 52,000 expenses were cleared</a>, with 35,000 of them being processed by primary hosts. Notably, OCF handles roughly half of all expenses on the platform.</p>
                </div>
            </details>

            <details>
                <summary>What are the host fees for various hosts?</summary>
                <div class="info-div-container">
                    <p><a href="#collective-expenses-by-hostfee">Fees range from 0% to 10%</a>, with many independent hosts not charging any fee.</p>
                </div>
            </details>

            <details>
                <summary>How much funding is routed through the platform? How is it distributed among hosts?</summary>
                <div class="info-div-container">
                    <p><a href="#collective-income-by-host-and-method">Close to 55 million USD transited the platform in the past year</a>, with approximately 39 million being managed by one of the three primary hosts.</p>
                </div>
            </details>

            <details>
                <summary>Where do the majority of funds originate? Through platform donations or top-ups by hosts? How does this distribution vary among hosts?</summary>
                <div class="info-div-container">
                    <p><a href="#collective-income-by-host-and-method">The bulk of the funding is added by hosts, not direct platform contributions</a>. The Open Source Collective stands out, with slightly more direct contributions than top-ups by hosts. Some prominent hosts, like numfocus, have a negligible amount of direct donations.</p>
                </div>
            </details>

            <details>
                <summary>What revenue is generated by platform tips? How does it differ across hosts and collectives?</summary>
                <div class="info-div-container">
                    <p>Platform tips arise solely from direct contributions. As highlighted earlier, most of these tips are directed to Open Source Collective. Nevertheless, they've chosen a higher host fee over platform tips. <a href="#collective-tips">Of the 330,000 USD accumulated as platform tips in the past year</a>, roughly 188,000 USD was attributed to Open Collective Foundation.</p>
                </div>
            </details>

            <details>
                <summary>Is Open Collective primarily a crowdfunding platform or a financial management platform?</summary>
                <div class="info-div-container">
                    <p>Only <a href="#crowdfunding">about 20% of the funds channelled through Open Collective are sourced from crowdfunding</a>. In my view, Open Collective has predominantly evolved into a financial management tool, albeit with crowdfunding features. A counterpoint suggests that our crowdfunding offerings haven't been updated for a while, indicating potential areas of enhancement.</p>
                </div>
            </details>

            <p>This report was compiled in September 2023, preceding the Open Collective team retreat in Bath. Authored by Hugi Asgeirsson. The source code to generate this report can be accessed <a href="https://github.com/opencollective/profile-report">here</a>.</p>
        """
    },
    'active_users': {
        'id': 'active-users',
        'title': 'Active Users',
        'figs': fig_dict['active_users'],
        'body': """
            <p>In this section we start by looking at the number of users on the platform in the last 12 months. There are two main definitions of "user" in this report:</p>

            <p><b>All Users:</b> Anyone who's done something on our platform, even just logged in, is included in this group.</p>    

            <p><b>Users with Counted Activities:</b> A subset of users who have performed at least 1 of a number of activities in the last 12 months. These are specific actions users do, grouped into different types:</p>

            <ul>
                <li><b>Expense Activities:</b> This includes creating expenses (collective.expense.created) and using virtual cards (virtualcard.purchase).</li>
                <li><b>Host Admin Activities:</b> These are tasks by host admins. They cover a range from approving a collective to managing virtual cards and paying expenses.</li>
                <li><b>Collective Admin Activities:</b> These are actions by collective admins like applying for collective status or editing members.</li>
                <li><b>Contributor Activities:</b> We don't pull these from the activity log. Instead, we focus on user actions, like making or updating an order. We look at user-updated orders marked as "PAID", "ACTIVE", or "CANCELLED".</li>
            </ul>

            <p>Some activities aren't included because they don't tie to a specific user group, like rejecting an expense or commenting on one.</p>
                            
            <p>In the rest of this report we will only be considering the "Users with Counted Activities" group.</p>

        """
    },
    'activity_counts': {
        'id': 'user-activity-counts',
        'title': 'User Activity',
        'figs': fig_dict['activity_counts'],
        'body': """
            <p>Here we look at the number of users who have performed these counted activities in the last 12 months. In the pie charts below we show the fraction of users who have performed each activity as a percentage of all users who have performed any "counted" activity. In the bar chart "histograms" we count the number of users who have performed the activity X number of times.</p>
            <p>Total Expenses is an aggregate of "Expense Activities", "Expense To Own Collective Activities" and "Virtual Card Purchases". Total Contributions is an aggregate of all contributor activities.</p>
        """
    },
    'user_cluster_intro': {
        'id': 'user-cluster-intro',
        'title': 'User Clustering',
        'figs': fig_dict['user_clusters_tree'],
        'body': """
            <p>Next, we group our users based on their typical activities on the platform. By doing this, we gain a clearer understanding of the diverse user categories we cater to on our platform.</p>

            <p>For instance, our data indicates that nearly 9% of users regularly submit expenses, while about 2% act on behalf of organizations. But, does a set of users engage in both activities? If so, is this subset significant? Manually exploring every potential activity combination would be cumbersome. This is where clustering proves invaluable. Clustering groups users by similarities without us making excessive assumptions. The aim is to discern patterns and identify distinct user types.</p>

            <p>From our analysis, 13 user groups emerged via clustering, each labeled by their defining characteristics. Delve into each cluster to see how users score on various activity metrics. A score of 1 in the "Collective admin" metric means that particular user tops the chart in collective admin activities platform-wide. Notably, these scores are logarithmic, ranging from 0 to 1. This scale enables us to equitably compare highly active users with their less active counterparts. For instance, a collective admin who approves 19 expenses annually might have a score of 0.47, while one that approves 190 expenses could score 0.85. Without such a scoring system, graphs would be less insightful due to the vast activity gap between the most active and the occasional users.</p>
        """
    },
    'user_clusters': {
        'id': 'user-clusters',
        'title': 'User Clusters',
        'figs': fig_dict['user_clusters'],
        'body': """"""
    },
    'user_clusters_tsne': {
        'id': 'user-clusters--tsne-plot',
        'title': 'User Clusters: TSNE Plot',
        'figs': fig_dict['tsne_result'],
        'body': """
            <p>Here we show a plot of the user clusters. The plot is interactive, so you can zoom in and hover over the points to see the user cluster labels.</p>
        """
    },
    'user_profiles': {
        'id': 'user-profiles',
        'title': 'User Profiles',
        'figs': fig_dict['user_profiling'],
        'body': """
            <p>Here we look at the user profiles for each of the user clusters. We show the top 8 users in some selected clusters, based on the cluster's defining characteristics. This helps us get a grasp of who the users are.</p>
        """
    },
    'collective_analysis': {
        'id': 'collective-analysis',
        'title': 'Collective Analysis',
        'body': """
            <p>Next we zoom in on collectives and their expenses and income. We will pay particular attention to relationships between money flow, host fee and platform tips.</p>

            <p>To create the graphs belew we needed to preprocess the data. In lay terms, think of this preprocessing as cleaning and organizing a messy room. Unnecessary items (like zero amounts or spammers) were thrown out, similar items (like transactions of the same collective) were grouped together, everything was labeled in a consistent manner, and some items (like internal transactions) were kept aside as they weren't relevant to the main analysis. The end goal is to have a clean and organized dataset that's easier to work with and understand.</p>

            <p>We explain in short here what steps were taken:</p>

            <ul>
                <li>For both expenses and income data, if the transaction wasn't already in USD, it was converted using a fixed exchange rate. This standardizes the currency across all transactions, making comparisons and aggregations straightforward.</li>
                <li>Any collective in the expenses data that had fewer than 5 transactions and was its own host was removed. Additionally, any collectives in the income data matching these dropped collectives were also removed. This was done to remove noise of people 'trying out' the platform and adding contributions and expenses to test hosts and collectives.</li>
                <li>In the income data, if the total number of transactions for a collective was less than 5, and if all payment methods for that collective were of type 'host', then these rows were removed. This is done likely to remove any noise or irrelevant data.</li>
                <li>Rows in both expenses and income datasets having a specific pattern in their slug (ending with <string>-<number>.<number>) were removed. This is because this pattern was identified as belonging to spammers.</li>
                <li>Any transactions in both datasets where the host collective was 'opencollective' were removed. This filters out internal payments to the OC Inc team.</li>
            </ul>

            <p>This resulted in a dataset of all expenses and contributions to all real collectives on the platform for the last 12 months.</p>
        """ 
    },
    'collective_analysis_expenses_by_host_and_collective': {
        'id': 'collective-expenses',
        'title': 'Collective Expenses',
        'figs': [fig_dict['treemaps']['expense_amount_by_host_and_collective'], fig_dict['treemaps']['expense_count_by_host_and_collective']],
        'body': """
            <p>In these first two graphs we look at the total amount and count of expenses by host and collective. We look at all hosts and collectives in a chart where the size of the square is proportional to its value.</p>
        """
    },
    
    'collective_analysis_expenses_by_hostfee': {
        'id': 'collective-expenses-by-hostfee',
        'title': 'Expenses by Host Fee',
        'figs': [fig_dict['treemaps']['expense_amount_by_host_fee'], fig_dict['treemaps']['expense_count_by_host_fee']],
        'body': """
            <p>Here we look at the total amount and count of expenses by host fee. We look at all hosts and collective in a chart where the size of the square is proportional to its value.</p>
            <p>Notice that in both cases, the zero-host-fee category is second largest by both counts and amount.</p>
        """
    },
    'collective_analysis_tip_potential': {
        'id': 'collective-income-by-host-and-method',
        'title': 'Tip Potential: Income by Host and Method',
        'figs': [fig_dict['treemaps']['income_amount_by_host_percentage_and_method']],
        'body': """
            <p>How large is the tip potential for each host? Here we look at the total amount of income by host and method. Tips are only possible for contributions, and are unlikely when organizations make larger donations via bank transfer.</p>
            <p>In this chart we categorize the income by host and method. We then color the methods according to their potential for tipping. By doing so we are investiging, particularly for the zero-host-fee hosts, what the potential for tipping is when all contributions are considered.</p>
        """
    },
    'collective_analysis_tip_income': {
        'id': 'collective-tips',
        'title': 'Tip Income: Tip Income by Host and Collective',
        'figs': [fig_dict['treemaps']['tip_amount_amount_by_host_fee_and_host_and_collective']],
        'body': """
            <p>We now look at the total amount of tips by host and collective. We color the squares by if the collective is on a host with a host fee or not. Hover over boxes to see tip income amounts.</p>
            <p>We earn about 240,000 USD in tips from hosts with a host fee, and about 91,000 USD from hosts without a host fee.</p>
        """
    },
    'collective_analysis_crowdfunding': {
        'id': 'crowdfunding',
        'title': 'Crowdfunding Income: Income by Host and Method',
        'figs': [fig_dict['treemaps']['income_crowdfunding']],
        'body': """
            <p>Is Open Collective a crowdfunding platform with money management features, or a money management platform with crowdfunding features? Let's have a look.</p>
            <p>Of the roughly 60M that went through the platform in the last 12 months, about 12M was from crowdfunding. That is about 20% of the total revenue.</p>
            <p>Open Source Collective stands out as the largest crowdfunding host, and the only major host with more income from crowdfunding than from other sources.</p>
        """
    }
}

# Generate HTML sections
intro_html_div = get_html_section(sections['intro'])
active_users_html_div = get_html_section(sections['active_users'], include_figs=True)
activity_counts_html_divs = get_html_figs_section(sections['activity_counts']['figs'], sections['activity_counts'])
user_clustering_intro_html_div = get_html_section(sections['user_cluster_intro'], include_figs=True, fig_height=600)
user_clusters_html_divs = get_html_figs_section(sections['user_clusters']['figs'], sections['user_clusters'])
tsne_html = get_html_figs_section(sections['user_clusters_tsne']['figs'], sections['user_clusters_tsne'], fig_height=1000)
user_profiles_html_divs = get_html_figs_section(sections['user_profiles']['figs'], sections['user_profiles'], fig_height=1400)
collective_analysis_html_divs = get_html_section(sections['collective_analysis'])
expense_amount_and_count_by_host_html = get_html_section(sections['collective_analysis_expenses_by_host_and_collective'], header_level='h2',include_figs=True, fig_height=600)
expense_amount_and_count_by_host_fee_html = get_html_section(sections['collective_analysis_expenses_by_hostfee'], header_level='h2',include_figs=True, fig_height=600)
income_by_host_and_method_html = get_html_section(sections['collective_analysis_tip_potential'], include_figs=True, header_level='h2',fig_height=600)
tip_amount_amount_by_host_fee_and_host_and_collective_html = get_html_section(sections['collective_analysis_tip_income'], include_figs=True, header_level='h2',fig_height=600)
income_crowdfunding_html = get_html_section(sections['collective_analysis_crowdfunding'], include_figs=True, header_level='h2',fig_height=600)


# Final HTML
final_html = f"""
<html>
    <meta charset="UTF-8">
    <head>
        <title>Retreat Report: Users, Hosts and Collectives</title>
        <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
        {css_styles}
    </head>

    <body>

        {generate_toc(sections)}

        <div style="margin: 0 70px 0 350px;">

            {intro_html_div}

            {active_users_html_div}

            {activity_counts_html_divs}

            {user_clustering_intro_html_div}

            {user_clusters_html_divs}

            {tsne_html}

            {user_profiles_html_divs}

            {collective_analysis_html_divs}

            {expense_amount_and_count_by_host_html}

            {expense_amount_and_count_by_host_fee_html}

            {income_by_host_and_method_html}

            {tip_amount_amount_by_host_fee_and_host_and_collective_html}

            {income_crowdfunding_html}

        </div>

    </body>
</html>
"""

# Write the final HTML to a file
with open("retreat_report.html", "w") as f:
    f.write(final_html)
